UPSTREAM: net: Move the permission check in sock_diag_put_filterinfo to packet_diag_dump
[ Upstream commit a53b72c83a4216f2eb883ed45a0cbce014b8e62d ]

The permission check in sock_diag_put_filterinfo is wrong, and it is so removed
from it's sources it is not clear why it is wrong.  Move the computation
into packet_diag_dump and pass a bool of the result into sock_diag_filterinfo.

This does not yet correct the capability check but instead simply moves it to make
it clear what is going on.

Reported-by: Andy Lutomirski <luto@amacapital.net>
Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from v3.14.9 commit 286a056788f863a2abe103a0efa46f6c412eb850)

TEST=cbuildbot + manual boot
BUG=chromium:427008

Change-Id: Iada559f9118c6790f48004fd87e79fc9dd4d4f1c
Signed-off-by: Kevin Cernekee <cernekee@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/244299
Reviewed-by: Kees Cook <keescook@chromium.org>

Conflicts:
	net/packet/diag.c
[filbranden: 3.18 rebase. Dropped commit, introduced upstream on:
 a53b72c83a42 net: Move the permission check in sock_diag_put_filterinfo to packet_diag_dump]
Signed-off-by: Filipe Brandenburger <filbranden@chromium.org>