CHERRY-PICK: CHROMIUM: iwl7000: don't disable SCD chain extension on newer devices
7000 device series have a fix for this hardware feature.
Stop disabling it, and get an improvement in Tx throughput.
This feature allows the scheduler to fetch more frames on
the fly while an A-MPDU is being built - which means that
we can get larger A-MPDU. This, of course, give an
improvement in the Tx throughput.

BUG=lower than expected TX throughput
TEST=run TX throughput test (preferably conductive) and check
throughput and aggregation sizes

Original-Change-Id: I2290250af5daaa92fb59bc17e3e17fe0226ad616
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/196798
Reviewed-by: Paul Stewart <pstew@chromium.org>
(cherry picked from commit 77c7b6ef9157bcf7fa3e23451ce4ed19c55fdecd)

Change-Id: Ib4241eac7204122d4ac8e431cbb9f56e7de3d6c2
Reviewed-on: https://chromium-review.googlesource.com/201530
Tested-by: Wei Shun Chang <wei.shun.chang@intel.com>
Reviewed-by: Jay Kim <yongjaek@chromium.org>
Commit-Queue: Wei Shun Chang <wei.shun.chang@intel.com>