UPSTREAM: [media] v4l2-ctrls: use ptrs for all but the s32 type
Rather than having two unions for all types just keep 'val' and
'cur.val' and use the p_cur and p_new unions to access all others.

The only reason for keeping 'val' and 'cur.val' is that it is used
all over, so converting this as well would be a huge job.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>
Signed-off-by: Mauro Carvalho Chehab <m.chehab@samsung.com>

BUG=chrome-os-partner:33728
TEST=build
Change-Id: I53bbde779169220e48236e24989b0190d1bc94c2
Reviewed-on: https://chromium-review.googlesource.com/225391
Reviewed-by: Wu-Cheng Li <wuchengli@chromium.org>
Commit-Queue: Pawel Osciak <posciak@chromium.org>
Tested-by: Pawel Osciak <posciak@chromium.org>