CHROMIUM: video: tegra: Add unplugged callback to hdmi.c
Add an unplugged callback that is called as soon as the state machine
determines that the output has been unplugged.  This will allow the
HDA block to be informed that there isn't an output any more and audio
can switch immediately.  Currently disabling the display path takes
over ten seconds because of some i2c timeouts and sleeps in nvhdcp,
avoiding that will let the user get his audio back faster.

BUG=chrome-os-partner:28420
TEST=plug/unplug HDMI listen for quicker audio switch now happens
after the 1.5 second de-bounce instead of after 12 seconds.

Change-Id: Ie311e79288ae8bc5d5b555f1754d940a886f15db
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/197566
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>