UPSTREAM: arm64: split syscall_trace() into separate functions for enter/exit
As done in arm, this change makes it easy to confirm we invoke syscall
related hooks, including syscall tracepoint, audit and seccomp which would
be implemented later, in correct order. That is, undoing operations in the
opposite order on exit that they were done on entry.

Acked-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: AKASHI Takahiro <takahiro.akashi@linaro.org>
Signed-off-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>
(cherry picked from commit 3157858feff89196635b01495d5ec9ebe206639e)
Signed-off-by: David Riley <davidriley@chromium.org>

Change-Id: Ie6215e68523c2705a98a95508101319f763ea0ac
Reviewed-on: https://chromium-review.googlesource.com/206205
Reviewed-by: Olof Johansson <olofj@chromium.org>
Tested-by: David Riley <davidriley@chromium.org>