UPSTREAM: net: cdc_ncm: fix buffer overflow
Commit 4d619f625a60 ("net: cdc_ncm: no point in filling up the NTBs
if we send ZLPs") changed the padding logic for devices with the ZLP
flag set.  This meant that frames of any size will be sent without
additional padding, except for the single byte added if the size is
a multiple of the USB packet size. But if the unpadded size is
identical to the maximum frame size, and the maximum size is a
multiplum of the USB packet size, then this one-byte padding will
overflow the buffer.

Prevent padding if already at maximum frame size, letting usbnet
transmit a ZLP instead in this case.

Fixes: 4d619f625a60 ("net: cdc_ncm: no point in filling up the NTBs if we send ZLPs")
Reported by: Yu-an Shih <yshih@nvidia.com>
Signed-off-by: Bjørn Mork <bjorn@mork.no>
Signed-off-by: David S. Miller <davem@davemloft.net>

(cherry picked from commit 9becd707841207652449a8dfd90fe9c476d88546)
Signed-off-by: Ben Chan <benchan@chromium.org>

BUG=chromium:419242
TEST=`USE=mbim emerge-$BOARD chromeos-kernel-3_14`
TEST=Run network3g tests on DUTs with a MBIM modem.

Change-Id: I3da1b20524bed8e01697a8b1acb78b49ee803776
Reviewed-on: https://chromium-review.googlesource.com/220704
Reviewed-by: Olof Johansson <olofj@chromium.org>
Commit-Queue: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>