ARC: Fix/unify cpu private IRQ requests (TIMER/IPI)
IPI IRQ request was broken as it was effectively being enabled on
boot-core only. The band aid was to make it like timer irq registration:
     cpu == 0 ? setup_irq() : arch_unmask_irq()

This is ugly (even in current timer code) as it requires need to expose
arch_unmask_irq() outside of intc code. So switch to percpu IRQ APIs:
  -request_percpu_irq [boot core]
  -enable_percpu_irq  [all cores]

Since TIMER and IPI both use this infrastructure, encapsulate this in a
helper arc_request_percpu_irq()

Signed-off-by: Vineet Gupta <vgupta@synopsys.com>