CHROMIUM: security: tlk_driver: Run kthreads on CPU0
Currently, the driver assumes that smc requests come either
from workqueue or from a thread that has possibility to set
cpu affinity. However, kernel threads by default do not have
this possibility and in principle the flag can be set in middle
of operation.

Turning on devices is sometimes done in kernel threads and in GPU
case we may need to change i.e. the VPR parameters. Therefore,
ensure that the communication is done on CPU0.

BUG=chrome-os-partner:42905
TEST=Tested SMC calls using TLK testapp

Change-Id: I0cbd8b841a07c8400310167fa8800eaae9de21f2
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Signed-off-by: Mahesh Lagadapati <mlagadapati@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/287708
Commit-Ready: David Riley <davidriley@chromium.org>
Tested-by: David Riley <davidriley@chromium.org>
Reviewed-by: David Riley <davidriley@chromium.org>