ipc,msg: loosen check for full queue
When sending a message, we must guarantee that there will be enough room
in the queue to add it, otherwise wait until space becomes available --
which requires blocking the calling task.  Currently, this relies meeting
both of the following conditions:

(i)  The new msg size + current size is lower than the queue's max size.

(ii) The current amount of messages in the queue + 1 (this msg) is
     lower than the queue's max size.

While (i) is obvious and well documented in the msgsnd(2) manpage, the
second one is far more ambiguous and does not serve the purpose, as we do
not control the amount of messages in the queue (unlike posix queues do).
Thus remove (ii) and loosen how we check for space.

Signed-off-by: Davidlohr Bueso <davidlohr@hp.com>
Cc: Manfred Spraul <manfred@colorfullife.com>
Cc: Aswin Chandramouleeswaran <aswin@hp.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>