UPSTREAM: Documentation: dt-bindings: update xhci-platform DT binding
This commit extends the compatible string list of the xhci-platform
binding with the new "armada-375-xhci" and "armada-380-xhci"
compatible strings. It is used to describe the XHCI controller which
is available in the Armada 375 and 38x SoCs.

It also indicates that an optional 'clocks' property is now supported.

Signed-off-by: Gregory CLEMENT <gregory.clement@free-electrons.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Acked-by: Mathias Nyman <mathias.nyman@linux.intel.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 023bfe83e278dca17abc491f8e1182211edb0559)
Signed-off-by: Benson Leung <bleung@chromium.org>

BUG=chrome-os-partner:30796
TEST=build tested for rush_ryu.

Change-Id: I120627b93a41c616a41c942ce1fc6d7847d45e82
Reviewed-on: https://chromium-review.googlesource.com/225790
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Tested-by: Benson Leung <bleung@chromium.org>
Commit-Queue: Benson Leung <bleung@chromium.org>