drivers: usb: hbm: Disable async scheduler when init a pipe
When we touch the HBM we need to make sure that the async
scheduler is not running. Otherwise it might stuck the controller
sooner or later. Since the controller uses the BAM for direct access
if anything gets stuck with the HBM, it might get the whole
controller unresponsive. (even endpoints unrelated to HBM, like
the control endpoint).
This change also adds the correct reset sequence steps
to the HBM pipes.

CRs-fixed: 613557, 591930
Change-Id: I13069909aafd682f52178f11debb95af45be28f4
Signed-off-by: Lena Salman <esalman@codeaurora.org>
Signed-off-by: Ido Shayevitz <idos@codeaurora.org>