CHROMIUM: wireless: Use correct Kconfig files selected by WIFIVERSION var
When WIFIVERSION=-3.8, we have been using the 3.14 Kconfig files, causing
issues such as missing iwl7000 driver. This patch ensures the correct
Kconfig files are selected according to WIFIVERSION.

I missed this during the 3.14 rebase because the change is in:
CHROMIUM: wireless-3.4: Make it build when WIFIVERSION=-3.4 is set
but not in:
CHROMIUM: wireless-3.8: Make it build when WIFIVERSION=-3.8 is set
and we dropped the 3.4wifi topic branch for 3.14.

BUG=none
TEST=run "WIFIVERSION=-3.8 chromeos/scripts/kernelconfig editconfig" and
see config options are coming from drivers/net/wireless-3.8/Kconfig

Change-Id: I62b48f70b9c7cda08091f9faafa79056b79530ba
Signed-off-by: Ben Zhang <benzh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/202550
Reviewed-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Sameer Nanda <snanda@chromium.org>