CHROMIUM: input: standardize elants_i2c sysfs
Currently the elants_i2c driver looks for it's new FW at
/lib/firmware/ElanFW.fw, but all our other touch drivers looks for a
file of the form /lib/firmware/$DRIVERNAME.bin, so this patch
changes the name to elants_i2c.bin for consistency and since that's
what the FW updating scripts expect.

The autoupdater also needs to have a hw_version sysfs entry where it can
read a number that identifies which device it's updating so it knows it
has the right one.  This patch adds that.

Additionally, the sysfs entry to update the FW used to be accessed by
reading from it, but all of our others are triggered by writing to it,
so this patch simple flips the functionality for the fw_update sysfs
entry when it moves them out of their nested group.

BUG=chrome-os-partner:29412
TEST=manual testing

Change-Id: I7a1286f00f9114b769df534fac23cd8edd615e47
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/202576
Reviewed-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Dennis Kempin <denniskempin@chromium.org>
(cherry picked from commit 213d2a6b64e310894172cc34a9d48056bd2c203f)
Reviewed-on: https://chromium-review.googlesource.com/203647