CHROMIUM: ARM: tegra: improve panel power-off time handling
On Big and Blaze the panel requires 500ms between panel power-off
and power-on, causing a noticeable delay in the power-off path.
Instead of sleeping for that amount of time unconditionally,
record the previous power-off time and use that to determine if
we need to sleep for any longer to meet the minimum power-off
requirement in the power-on path.

BUG=chrome-os-partner:28128
TEST=Boot on Big, suspend/resume, display still works.
Time to enter suspend is reduced.

Change-Id: I548228ad5a0a4428bfaa6b6cfe04f7b2e8786472
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/202578
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Mark Zhang <markz@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/203811