BACKPORT: drm/tegra: sor - Don't hardcode link parameters
The currently hardcoded link parameters don't work on all eDP panels, so
compute the parameters at runtime depending on the mode and panel type
to allow the driver to cope with a wider variety of panels.

Note that the number of bits per pixel of the panel is still hardcoded,
but this can be addressed in a separate patch.

This is largely based on a patch by Stéphane Marchesin but the algorithm
was largely rewritten to be more readable and concise.

Signed-off-by: Stéphane Marchesin <marcheu@chromium.org>
Signed-off-by: Thierry Reding <treding@nvidia.com>
(cherry picked from commit 34fa183bacf9b5ecfda864857e8a797065b6b7e8)
Signed-off-by: Stéphane Marchesin <marcheu@chromium.org>

BUG=chromium:385270
TEST=compiles

Change-Id: I7859c82e0a57adf2bb339f46fc71e0fc90a1af29
Reviewed-on: https://chromium-review.googlesource.com/208438
Reviewed-by: Olof Johansson <olofj@chromium.org>
Tested-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Stéphane Marchesin <marcheu@chromium.org>