CHERRY-PICK: Bluetooth: Reuse hci_stop_discovery function when cleaning up HCI state
When cleaning up the HCI state as part of the power-off procedure we can
reuse the hci_stop_discovery() function instead of explicitly sending
HCI command related to discovery. The added benefit of this is that it
takes care of canceling name resolving and inquiry which were not
previously covered by the code.

BUG=chrome-os-partner:28815
TEST=Enable and disable the BT device multiple times

Signed-off-by: Johan Hedberg <johan.hedberg@intel.com>
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>
(cherry picked from commit 986603b8949d4347717105556f774f956d8fc605)
Signed-off-by: Tedd Ho-Jeong An <tedd.an@intel.com>

Change-Id: Ie7c62c949e5899cb06323c164258b76e1971dc4b
Reviewed-on: https://chromium-review.googlesource.com/203788
Reviewed-by: Scott James Remnant <keybuk@chromium.org>
Commit-Queue: Tedd An <tedd.an@intel.com>
Tested-by: Tedd An <tedd.an@intel.com>