CHROMIUM: mfd: tps65090: Fix compile error.
The following two commits are incompatible and a result of the
3.14 rebase.  The former constifies a structure modified in the
latter patch.  This is fixed upstream (as part of a larger set
of changes that rely on some other changes) and this driver probably
won't be enabled in a product shipping on 3.14, so just make things
compile without backporting all the other changes.

The incompatible changes:
    mfd: ti: Constify struct mfd_cell where possible
    mfd: tps65090: Allow charger module to be used when no irq

BUG=None
TEST=compiles

Change-Id: Ib48cab0ab7df73bc4fae8a3f1059fa2c12f0d149
Signed-off-by: David Riley <davidriley@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/206737
Reviewed-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>