ARM: tegra: move fuse exports to tegra-soc.h
All fuse related functionality will move to a driver in the following
patches. To prepare for this, export all the required functionality in a
global header file and move all users of fuse.h to tegra-soc.h. While
we're at it, remove tegra_bct_strapping, as its only user was removed in
Commit a7cbe92cef27 ("ARM: tegra: remove tegra EMC scaling driver").

Signed-off-by: Peter De Schrijver <pdeschrijver@nvidia.com>
Signed-off-by: Stephen Warren <swarren@nvidia.com>
Signed-off-by: Thierry Reding <treding@nvidia.com>