UPSTREAM: spi: qup: Remove chip select function
This patch removes the chip select function.  Chip select should instead be
supported using GPIOs, defining the DT entry "cs-gpios", and letting the SPI
core assert/deassert the chip select as it sees fit.

The chip select control inside the controller is buggy.  It is supposed to
automatically assert the chip select based on the activity in the controller,
but it is buggy and doesn't work at all.  So instead we elect to use GPIOs.

Signed-off-by: Andy Gross <agross@codeaurora.org>
Signed-off-by: Mark Brown <broonie@linaro.org>
(cherry picked from commit 4a8573abe965115bc5b064401fd669b74e985258)
BUG=chrome-os-partner:31587
TEST=build/boot on AP-148 and storm P0

Change-Id: Ieee85476ae4e90b6a6ac726703a155d6f9d078bb
Signed-off-by: Grant Grundler <grundler@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/215017
Reviewed-by: Olof Johansson <olofj@chromium.org>