mm/pagewalk: replace mm_walk->skip with more general mm_walk->control
Originally mm_walk->skip is used to determine whether we walk over a vma
or not.  But this is not enough because one of the page table walker's
caller subpage_mark_vma_nohuge(), will need another behavior PTWALK_BREAK,
which let us break current loop and continue from the beginning of the
next loop.

To implement this behavior and make it extensible for future users, this
patch replaces mm_walk->skip with more flexible mm_walk->control, and
changes its default value to PTWALK_NEXT (which is equivalent to present
walk->skip == 1.) This is because PTWALK_NEXT provides the behavior which
is most likely to be used globally.

Signed-off-by: Naoya Horiguchi <n-horiguchi@ah.jp.nec.com>
Cc: Dave Hansen <dave.hansen@intel.com>
Cc: Hugh Dickins <hughd@google.com>
Cc: "Kirill A. Shutemov" <kirill@shutemov.name>
Cc: Johannes Weiner <hannes@cmpxchg.org>
Cc: Michal Hocko <mhocko@suse.cz>
Cc: KOSAKI Motohiro <kosaki.motohiro@jp.fujitsu.com>
Cc: KAMEZAWA Hiroyuki <kamezawa.hiroyu@jp.fujitsu.com>
Cc: "Aneesh Kumar K.V" <aneesh.kumar@linux.vnet.ibm.com>
Cc: Pavel Emelyanov <xemul@parallels.com>
Cc: Rik van Riel <riel@redhat.com>
Cc: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>