UPSTREAM: kernel: time: Add udelay_test module to validate udelay
Create a module that allows udelay() to be executed to ensure that
it is delaying at least as long as requested (with a little bit of
error allowed).

There are some configurations which don't have reliably udelay
due to using a loop delay with cpufreq changes which should use
a counter time based delay instead.  This test aims to identify
those configurations where timing is unreliable.

Signed-off-by: David Riley <davidriley@chromium.org>
Signed-off-by: John Stultz <john.stultz@linaro.org>
(cherry picked from git://git.linaro.org/people/jstultz/linux
 commit 5085032326671516816c27add114ad0ef69a023d)

BUG=chromium:352391,chromium:395341
TEST=no more warning when building for MIPS

Change-Id: I9bd5ffa261c940817175eb2fd33c2fcdd044ee7f
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/209375
Reviewed-by: Olof Johansson <olofj@chromium.org>
Reviewed-by: David Riley <davidriley@chromium.org>