CHROMIUM: pinctrl: tegra-xusb: Add USB PHY support
In addition to the PCIe and SATA PHYs, the XUSB pad controller also
supports 3 UTMI, 2 HSIC, and 2 USB3 PHYs.  Each USB3 PHY uses a single
PCIe or SATA lane and is mapped to one of the three UTMI ports.

The xHCI controller will also send messages intended for the PHY driver,
so request and listen for messages on the mailbox's PHY channel.

Signed-off-by: Thierry Reding <treding@nvidia.com>
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>

BUG=chrome-os-partner:38876
TEST=Build and boot on Smaug.

Change-Id: I4eade2450af2ab53bdb6ea7e77d259a63c489b47
Reviewed-on: https://chromium-review.googlesource.com/269248
Reviewed-by: Benson Leung <bleung@chromium.org>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>
Tested-by: Andrew Bresticker <abrestic@chromium.org>