FROMLIST: usb: xhci: Add NVIDIA Tegra xHCI host-controller driver
Add support for the on-chip xHCI host controller present on Tegra SoCs.

The driver is currently very basic: it loads the controller with its
firmware, starts the controller, and is able to service messages sent
by the controller's firmware.  The hardware also supports device mode
as well as powergating of the SuperSpeed and host-controller logic
when not in use, but support for these is not yet implemented.

Based on work by:
  Ajay Gupta <ajayg@nvidia.com>
  Bharath Yadav <byadav@nvidia.com>

Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-by: Stephen Warren <swarren@nvidia.com>
[bleung : fixup Kconfig]
(am from https://patchwork.kernel.org/patch/4926361/)
Signed-off-by: Benson Leung <bleung@chromium.org>

BUG=chrome-os-partner:30796
TEST=build tested for rush_ryu.

Change-Id: I0b46eab7e663fd56c6d8388f05e9631a3831bd10
Reviewed-on: https://chromium-review.googlesource.com/219615
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Tested-by: Benson Leung <bleung@chromium.org>
Commit-Queue: Benson Leung <bleung@chromium.org>