BACKPORT: seccomp: add "seccomp" syscall
This adds the new "seccomp" syscall with both an "operation" and "flags"
parameter for future expansion. The third argument is a pointer value,
used with the SECCOMP_SET_MODE_FILTER operation. Currently, flags must
be 0. This is functionally equivalent to prctl(PR_SET_SECCOMP, ...).

In addition to the TSYNC flag later in this patch series, there is a
non-zero chance that this syscall could be used for configuring a fixed
argument area for seccomp-tracer-aware processes to pass syscall arguments
in the future. Hence, the use of "seccomp" not simply "seccomp_add_filter"
for this syscall. Additionally, this syscall uses operation, flags,
and user pointer for arguments because strictly passing arguments via
a user pointer would mean seccomp itself would be unable to trivially
filter the seccomp syscall itself.

Signed-off-by: Kees Cook <keescook@chromium.org>
Reviewed-by: Oleg Nesterov <oleg@redhat.com>
Reviewed-by: Andy Lutomirski <luto@amacapital.net>

BUG=chromium:308775
TEST=rambi build, boot, seccomp test suite passes

Change-Id: I507a07fde3f58045cc611c651747606ce4e1935f
(cherry picked from commit 48dc92b9fc3926844257316e75ba11eb5c742b2c)
[stubbed out sys_renameat2]
Signed-off-by: Kees Cook <keescook@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/218969
Reviewed-by: Benson Leung <bleung@chromium.org>

Conflicts:
	arch/Kconfig
	arch/x86/syscalls/syscall_32.tbl
	arch/x86/syscalls/syscall_64.tbl
	include/linux/syscalls.h
	include/uapi/asm-generic/unistd.h
	include/uapi/linux/seccomp.h
	kernel/seccomp.c
	kernel/sys_ni.c
[filbranden: 3.18 rebase. Dropped commit, introduced upstream on:
 48dc92b9fc39 seccomp: add "seccomp" syscall]
Signed-off-by: Filipe Brandenburger <filbranden@chromium.org>