FROMLIST: of: Add NVIDIA Tegra memory controller binding
The memory controller on NVIDIA Tegra exposes various knobs that can be
used to tune the behaviour of the clients attached to it.

In addition, the memory controller implements an SMMU (IOMMU) which can
translate I/O virtual addresses to physical addresses for clients. This
is useful for scatter-gather operation on devices that don't support it
natively and for virtualization or process separation.

Changes in v3:
- add missing clocks and clock-names properties
- add example

BUG=chrome-os-partner:31213
TEST=boot with the rest of the DRM series, see full resolution graphics work

Change-Id: I79625ba36e4a892b7a6c627a79243d9134e868a4
Signed-off-by: Thierry Reding <treding@nvidia.com>
Signed-off-by: Olof Johansson <olofj@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/223569
Reviewed-by: David Riley <davidriley@chromium.org>