FROMLIST: clk: composite: support determine_rate using rate_ops->round_rate + mux_ops->set_parent
In case the rate_hw does not implement determine_rate, but only round_rate
we fallback to best_parent selection if mux_hw is present and support
reparenting.

Signed-off-by: Boris BREZILLON <b.brezillon@overkiz.com>

This also fixes a rate calculation problem when using the standard div and
mux ops, as in this case currently only the mux->determine_rate is used
in the composite rate calculation.
So when for example the composite clock has two parents at 600 and 800MHz,
the requested rate is 75MHz, which the divider could provide, without this
change the rate would be set 600MHz ignoring the divider completely.
This may be way out of spec for the component.

BUG=chrome-os-partner:30167
TEST=Used by future rockchip patches

[fixed the output to actually return a rate instead of the diff]
Change-Id: I8b2ce34ac618146f42884d2e6ce5de7516e40a66
Signed-off-by: Heiko Stuebner <heiko@sntech.de>
Acked-by: Max Schwarz <max.schwarz@online.de>
Tested-by: Max Schwarz <max.schwarz@online.de>
Tested-by: Gabriel Fernandez <gabriel.fernandez@linaro.org>
Signed-off-by: Doug Anderson <dianders@chromium.org>
(am from https://patchwork.kernel.org/patch/4468541/)
Reviewed-on: https://chromium-review.googlesource.com/207594
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>