FROMLIST: clk: rockchip: add clock driver for rk3188 and rk3066 clocks
This adds a clock driver that handles the specific muxes, dividers and gates
of rk3188 and rk3066 SoCs.

The structure of the clock list resembles the arrangement of their
counterparts in the clock architecture diagrams found in the SoC
documentation.

Clocks exported to the clock provider are currently limited to well known
or measured ones. So additional clock exports may be necessary in the future.

BUG=chrome-os-partner:30167
TEST=None

Change-Id: Ifd1c071fffb523ea2c66d34255d1f45294f0f030
Signed-off-by: Heiko Stuebner <heiko@sntech.de>
Acked-By: Max Schwarz <max.schwarz@online.de>
Tested-By: Max Schwarz <max.schwarz@online.de>
Signed-off-by: Doug Anderson <dianders@chromium.org>
(am from https://patchwork.kernel.org/patch/4468611/)
Reviewed-on: https://chromium-review.googlesource.com/207667
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>