CHROMIUM: irqchip: tegra: Add exception for T132
T132 advertises 6 controllers, but the 6th controller is unique
and needs to be special cased.

The sixth controller is the Hierarchical Interrupt Controller
and is different from the other 5 controllers on Tegra132. If we
need to add support for it, either this driver will need to be
extended, or we may need to create another driver similar to how
downstream handles it.

However, the Hierarchical Interrupt Controller is mainly for
debug purposes, so we likely don't need to worry about it.

BUG=chrome-os-partner:31211
TEST=IRQ controller comes up and interrupts are functional

Change-Id: I8ae8329f75284c251b896502d2d930449ccee525
Signed-off-by: Rhyland Klein <rklein@nvidia.com>
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/220044
Reviewed-by: David Riley <davidriley@chromium.org>
Signed-off-by: David Riley <davidriley@chromium.org>
Reviewed-on: https://chrome-internal-review.googlesource.com/209054
Reviewed-by: David Riley <davidriley@google.com>
Commit-Queue: David Riley <davidriley@google.com>
Tested-by: David Riley <davidriley@google.com>
Reviewed-on: https://chromium-review.googlesource.com/267264