UPSTREAM: clk: qcom: Support bypass RCG configuration
In the case of HDMI clocks, we want to bypass the RCG's ability
to divide the output clock and pass through the parent HDMI PLL
rate. Add a simple set of clk_ops to configure the RCG to do
this. This removes the need to keep adding more frequency entries
to the tv_src clock whenever we want to support a new rate.

Tested-by: Rob Clark <robdclark@gmail.com>
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>
(cherry picked from commit 404c1ff67d241e8503fd46a09d324343aa321a8a)
BUG=chrome-os-partner:31587
TEST=build/boot on AP-148 and storm P0

Change-Id: I942c90a09ed324ebecaa0d9ffe608647968a6381
Signed-off-by: Grant Grundler <grundler@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/214984
Reviewed-by: Olof Johansson <olofj@chromium.org>