UPSTREAM: HID: rmi: make compututation of the address of Query 12 more careful
There are additional queries which are optional and may not be present
depending on the configuration of the firmware. Knowing which queries are
present is needed to properly compute the address of Query 12 and all
subsequent queries. Additional bits in Query 1 are used to indicate the
presence of these optional queries.

Signed-off-by: Andrew Duggan <aduggan@synaptics.com>
Reviewed-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
(cherry picked from commit 9e2c327ee85e37264a1e04109efc7b97844afd0b)
Signed-off-by: Andrew Duggan <aduggan@synaptics.com>

BUG=chrome-os-partner:32116
TEST=With all patches in this series applied on retrofitted Pixel. Synaptics
touchpad is functional.

Change-Id: I98381b88c39e9f446f6c903eeef3f25a197d4862
Reviewed-on: https://chromium-review.googlesource.com/220086
Reviewed-by: Benson Leung <bleung@chromium.org>
Commit-Queue: Andrew Duggan <aduggan@synaptics.com>
Tested-by: Andrew Duggan <aduggan@synaptics.com>