ASoC: Add ADAU7002 Stereo PDM-to-I2S/TDM Converter driver
This patch adds support for the ADAU7002 PDM-to-I2S/TDM Converter. The ADAU7002
takes a stereo PDM signal (e.g. from two digital microphones) and converts it
into a I2S/TDM PCM stream. The chip does not have a programmable interface and
the driver describes the static capabilities of the chip.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>