FROMLIST: ARM: tegra: Add memory controller support for Tegra114
Add the device tree node for the memory controller found on Tegra114
SoCs. The memory controller integrates an IOMMU (called SMMU) as well as
various knobs to tweak memory accesses by the various clients.

The old IOMMU device tree node is collapsed into the memory controller
node to more accurately describe the hardware. While this change is
incompatible, the IOMMU driver has never had any users so the change is
not going to cause any breakage.

BUG=chrome-os-partner:31213
TEST=boot with the rest of the DRM series, see full resolution graphics work

Change-Id: I65aa3c8c514d016f73eee15c16134d0681b9e4ab
Signed-off-by: Thierry Reding <treding@nvidia.com>
Signed-off-by: Olof Johansson <olofj@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/223604
Reviewed-by: David Riley <davidriley@chromium.org>